requirements.txt
Django==5.2.8
django-admin-interface>=0.25.0
django-colorfield>=0.12.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
Pillow>=10.0.0
whitenoise>=6.6.0
